+++ src/Makefile.am
@@ -596,7 +596,6 @@ libdash_common_a_SOURCES = \
   key.cpp \
   key_io.cpp \
   keystore.cpp \
-  llmq/params.cpp \
   netaddress.cpp \
   netbase.cpp \
   net_permissions.cpp \
